+++ src/icons/checkmark-icon.tsx
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2011-2019 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+/**
+ * Check-mark icon
+ */
+export const CheckMarkIcon = createIcon(
+    <svg viewBox="0 0 20 20" fill="none">
+        <path
+            d="M5 9.6l3.5 3.5L15 6.6"
+            className="gs-svg-icon-stroke"
+            strokeWidth="2"
+            strokeLinecap="round"
+        />
+    </svg>
+);