+++ src/icons/icons.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2011-2019 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+
+export {ArrowDownloadIcon} from './arrow-download-icon';
+export {ArrowLeftIcon} from './arrow-left-icon';
+export {ArrowRightIcon} from './arrow-right-icon';
+export {BookmarkIcon} from './bookmark-icon';
+export {CheckMarkIcon} from './checkmark-icon';
+export {ClearIcon} from './clear-icon';
+export {CrossIcon} from './cross-icon';
+export {DownloadIcon} from './download-icon';
+export {EyeIcon} from './eye-icon';
+export {FileIcon} from './file-icon';
+export {HelpIcon} from './help-icon';
+export {InfoIcon} from './info-icon';
+export {KeyboardArrowBottomIcon} from './keyboard-arrow-bottom-icon';
+export {KeyboardArrowLeftIcon} from './keyboard-arrow-left-icon';
+export {KeyboardArrowRightIcon} from './keyboard-arrow-right-icon';
+export {LedIcon} from './led-icon';
+export {LockIcon} from './lock-icon';
+export {MenuIcon} from './menu-icon';
+export {MoreIcon} from './more-icon';
+export {OpenFolderIcon} from './open-folder-icon';
+export {PauseIcon} from './pause-icon';
+export {PlayCircledIcon} from './play-circled-icon';
+export {PlayIcon} from './play-icon';
+export {PlusIcon} from './plus-icon';
+export {PlusUserIcon} from './plus-user-icon';
+export {SearchIcon} from './search-icon';
+export {ShareIcon} from './share-icon';
+export {StopIcon} from './stop-icon';
+export {WarningIcon} from './warning-icon';