+++ src/icons/keyboard-arrow-left-icon.tsx
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2011-2019 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+/**
+ * Keyboard Left Icon (aka chevron turned left, aka >) icon
+ */
+export const KeyboardArrowLeftIcon = createIcon(
+    <svg viewBox="0 0 20 20">
+        <path
+            d="M12 6L8 10L12 14"
+            fill="none"
+            className="gs-svg-icon-stroke"
+            strokeWidth="1.5"
+            strokeLinecap="round"
+        />
+    </svg>
+);